fix: show calendar turbo_stream & add views switch button
