+++ app/views/layouts/_toast.html.erb
@@ -1,5 +1,7 @@
-<div class="absolute right-0 px-5 py-2 top-10">
-    <% [alert, notice, success, info, mail].each do |message| %>
-        <%= render partial: "shared/#{message}", locals: { message: }  if message %>
-    <% end %>
+<div class="absolute right-0 z-10 px-5 py-2 top-10">
+  <%= render partial: 'shared/alert', locals: { message: flash[:alert]} %>
+  <%= render partial: 'shared/notice', locals: { message: flash[:notice]} %>
+  <%= render partial: 'shared/success', locals: { message: flash[:success]} %>
+  <%= render partial: 'shared/info', locals: { message: flash[:info]} %>
+  <%= render partial: 'shared/mail', locals: { message: flash[:mail]} %>
 </div>
