+++ app/views/projects/_calendar.html.erb
@@ -1,6 +1,6 @@
 <div id="calendar"
-  data-controller="ccc"
-  data-ccc-target="calendar" 
+  data-controller="calendar"
+  data-calendar-target="calendar" 
   data-events="<%= events %>"
   data-new-event-url="<%= new_event_project_path %>"
   data-originalUrl="<%= request.original_url %>"
