+++ app/views/projects/kanban_view.html.erb
@@ -1,11 +1,5 @@
-<div id="project_show">
-  <%= @project.title %>
-  <%= @project.description %>
-  <%= @project.owner.nick_name %>
-  <%= link_to 'X', project_path(@project), data: { turbo_method: 'delete', turbo_confirm: '確定刪除專案嗎？'} %>
-  <%= link_to  edit_project_path, data: { turbo_frame: "modal" } do%>
-    <i class="fa-solid fa-ellipsis"></i>
-  <% end %>
+<div>
+  <%= render partial: 'head', locals: {project: @project} %>
 </div>
 <hr>
